fix: Remove auth callback page and use loader on login modal
